1. Remove Align256 function
1. Add a few new functions to IAllocator for allocating an array with safety check
1. Change the AllocatorMap in IExecutionProvider, add a device id as key
1. Replace the atomic_int in MLValueNameIdxMap, by a plain int.
1. Add an alignment argument to GetSizeInBytesFromTensorProto function
1. Fix warnings
1. Fix a bug in MKL-DNN. Allocator's memtype is wrong.

Before having this change, there is no way to support multiple GPU. As for each memory type, each provider can only have one allocator.

Related work items: #1745

+++ onnxruntime/contrib_ops/cpu/attnlstm/deep_cpu_attn_lstm.cc
@@ -167,7 +167,7 @@ Status DeepCpuAttnLstmOp::ComputeImpl(OpKernelContext& context) const {
   // output shape is [seq_length, num_directions, batch_size, hidden_size]
   // so it's not a case of all the output for one direction being first.
   // due to that we can only easily check that the end of the output for each direction is valid.
-  const size_t output_size = Y != nullptr ? Y->Shape().Size() : 0;
+  const size_t output_size = Y != nullptr ? static_cast<size_t>(Y->Shape().Size()) : 0;
   const size_t per_direction_offset = batch_size * hidden_size_;
   gsl::span<T> output = Y != nullptr ? Y->MutableDataAsSpan<T>() : gsl::span<T>();
   gsl::span<T> output_1 =
